[FEAT] #44 wrap NoticesPage Suspense
